Add role selection to onboarding flow
