+++ src/ui/types/index.ts
@@ -11,6 +11,7 @@ export type ExperimentalUserSettings = {
   disableLogRocket: boolean;
   enableTeams: boolean;
   enableLargeText: boolean;
+  role: string;
 };
 
 export type LocalExperimentalUserSettings = {
