+++ src/ui/utils/telemetry.ts
@@ -85,6 +85,7 @@ export type TelemetryUser = {
   id: string | undefined;
   email: string | undefined;
   internal: boolean;
+  role?: string;
 };
 
 let telemetryUser: TelemetryUser | undefined;
